Beginnings of "onion" tutorial
This is the start at sketching out the form of Antranig's "onion tutorial"
which will start very small and build step by step. Ultimately, the narrative
will move to the docs website, and be supplemented with links to related pages,
etc. The final form of these code files will likely change, as well.